refactor/refactor path directory files
