+++ src/components/Main/Main.js
@@ -1,12 +1,12 @@
 import React from "react";
 import "./Main.css";
 import Header from "../Header/Header";
-import Promo from "../Promo/Promo";
-import NavTab from "../NavTab/NavTab";
-import AboutProject from "../AboutProject/AboutProject";
-import Techs from "../Techs/Techs";
-import AboutMe from "../AboutMe/AboutMe";
-import Portfolio from "../Portfolio/Portfolio";
+import Promo from "./Promo/Promo";
+import NavTab from "./NavTab/NavTab";
+import AboutProject from "./AboutProject/AboutProject";
+import Techs from "./Techs/Techs";
+import AboutMe from "./AboutMe/AboutMe";
+import Portfolio from "./Portfolio/Portfolio";
 import Footer from "../Footer/Footer";
 
 function Main() {
