+++ src/components/Main/Portfolio/Portfolio.css
@@ -38,7 +38,7 @@
 
 .portfolio__link::after {
   content: "";
-  background-image: url(../../images/icon__portfolio.svg);
+  background-image: url(../../../images/icon__portfolio.svg);
   background-repeat: no-repeat;
   background-size: contain;
   background-position: center;
