+++ src/components/Main/Techs/Techs.css
