+++ src/components/Main/Techs/Techs.js
